+++ events/ready.js
@@ -5,24 +5,32 @@ const ownerID = process.env.OWNERID_ID
 const punishments = require("../models/ModSchema");
 const moment = require ("moment")
 const currenttime = moment(Date.now()).format('DD/MM/YY HH:mm:ss');
+const generic_server = process.env.GENERIC_SERVER
+const member_count = process.env.MEMBER_COUNT
+const ready_channel = process.env.READY_CHANNEL
 module.exports = {
 	name: 'ready',
-	async execute(client, messageCreate) {
-	const server = await client.guilds.cache.find(g => g.id == "864306095388229632");
+	once: true,
+
+	async execute(client) {
+	const server = await client.guilds.cache.get(generic_server);
+	if(!server){
+		return console.log(`Server not found`)
+	}
 	const owner = server.members.cache.get(ownerID)
 	try{
 	let guilds = client.guilds.cache.map(guild => "Guild name" +  "    " + "     " + guild.name + "    " + "     " + `Guild id` +  "    " + "     " + guild.id)
 	console.log(guilds)
-	const membercount = await server.channels.cache.find(ch => ch.id === '864306096004399148');
+	const membercount = await server.channels.cache.get(member_count);
 const discordJSVersion = packageJSON.dependencies["discord.js"];
 await server.members.fetch()
 let membernum = server.memberCount
     membercount.setName(`Members: ${membernum}`)
 client.user.setActivity(`Slava generic! My prefix is ${prefix}`)
 console.log(`Connected to discord, Successfully logged as ${client.user.username}, Runtime Enviorement Version (Node)${process.version} using discord.js version: ${discordJSVersion}`)
-await server.members.fetch(); // Fetch all members to ensure they're cached
+await server.members.fetch();
 
-const members = server.members.cache.filter(member => !member.user.bot); // Ignore bots
+const members = server.members.cache.filter(member => !member.user.bot);
 
 for (const member of members.values()) {
 	let data = await punishments.findOne({
@@ -36,7 +44,7 @@ for (const member of members.values()) {
 		await punishments.create({
 			GuildID: server.id,
 			UserID: member.id,
-			Punishments: [] // Initialize with an empty array
+			Punishments: []
 		});
 		console.log(`Initialized modlog for ${member.user.tag}`);
 	}
@@ -46,7 +54,7 @@ let readyembed = new Discord.EmbedBuilder()
 .setTitle(`Bot initialised and ready!`)
 .setColor("Green")
 .setFooter({text: `${currenttime}`})
-let readychannel = server.channels.cache.find(ch => ch.id == "1355718490673119355")
+let readychannel = server.channels.cache.get(ready_channel)
 readychannel.send({embeds: [readyembed]})
 		} catch(e){
 			console.log(e)