+++ CHANGELOG.md
@@ -1,12 +1,50 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [v1.1.0] - 
+
+## Added
+
+- Added censor command to add, remove or view censor list from mongodb
+
+- Added whitelist command to add, remove or view whitelist from mongodb
+
+- Added new whitelistSchema to models/
+
+- Added new censorSchema to models/
+
+- Added envexample.txt to provide an example of .env layout
+
+## Changed
+
+- Moved from local json file to customisable mongodb for censor and whitelist functions
+
+- moved THREAD_CATEGORY_ID, MEMBER_COUNT, READY_CHANNEl, ALERT_CHANNEL, GENERAL_CHAT, MEMBER_ROLE, SUGGESTIONS_CHANNEL, MUTE_ROLE to .env instead of being defined in commands
+
+- Channel ID's are no longer hardcoded in individual command files
+
+- Added installation instructions to README.md
+
+- Set mongoose strictQuery to false in index.js
+
+- changed how usernames of members are resolved in delete.js and modlogs.js
+
+- Improved error handling in messageCreate.js for modmail section. 
+
+- Allow attachments to be sent by staff and users in modmail
+
+-  Improved event logging
+## Removed
+- Removed censor.json & whitelist.json
+--
 
 ## [v1.0.1] - 27-05-25
 
 ### Added
 
 - Fixed up **train.js** for better code readability and end-user experience
+
+---
 
 ## [v1.0.0] - 2025-05-25